Mandate UTF-8 as file.encoding.
A better fix would avoid use of the platform default encoding
altogether.

Also adds a meta directive the the generated HTML to set the
encoding to UTF-8. We assume that the page that this is embedded
into follows the same conventions, either with a meta tag or with
a HTTP header.