twitter stuff is published now except for Finagle itself
at least, I think that's more or less accurate. happy to accept
corrections.

not all of the tickets got closed, but it seems to be because a few
random submodules (perhaps just for testing) weren't ported yet, that
kind of thing, rather than actual unavailability of the artifacts
people are likely to want